+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -9228,6 +9228,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         matching_item_rows_by_label: bool,
         index_is_bool_indexer: bool,
         deduplicate_columns: bool = False,
+        is_item_series: bool = False,
     ) -> "SnowflakeQueryCompiler":
         """
         Create a new SnowflakeQueryCompiler with indexed columns and rows replaced by item.
@@ -9252,6 +9253,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             index_is_bool_indexer: if True, the index is a boolean indexer.
             deduplicate_columns: if True, deduplicate columns from ``columns``, e.g., if columns = ["A","A"], only the
                 second "A" column will be used.
+            is_item_series: if item is from a Series
         Returns:
             Updated SnowflakeQueryCompiler
         """
@@ -9277,6 +9279,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             matching_item_rows_by_label=matching_item_rows_by_label,
             index_is_bool_indexer=index_is_bool_indexer,
             deduplicate_columns=deduplicate_columns,
+            item_is_series=is_item_series,
         )
 
         return SnowflakeQueryCompiler(result_frame)